+++ src/test/java/com/artipie/rpm/TestResource.java
@@ -21,45 +21,44 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artipie.rpm.files;
+package com.artipie.rpm;
 
-import com.artipie.rpm.TestResource;
-import java.io.IOException;
-import java.nio.file.Files;
+import java.net.URISyntaxException;
 import java.nio.file.Path;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.hamcrest.core.IsEqual;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
+import java.nio.file.Paths;
 
 /**
- * Test for {@link Gzip}.
- * @since 0.8
+ * Get file by name from test resources.
+ * @since 0.11
  */
-class GzipTest {
+public final class TestResource {
 
-    @Test
-    void unpacks(final @TempDir Path tmp) throws IOException {
-        new Gzip(new TestResource("test.tar.gz").file()).unpackTar(tmp);
-        MatcherAssert.assertThat(
-            Files.readAllLines(tmp.resolve("test.txt")).iterator().next(),
-            new IsEqual<>("hello world")
-        );
+    /**
+     * File name.
+     */
+    private final String name;
+
+    /**
+     * Ctor.
+     * @param name File name
+     */
+    public TestResource(final String name) {
+        this.name = name;
     }
 
-    @Test
-    void unpacksGz(final @TempDir Path tmp) throws IOException {
-        final Path res = tmp.resolve("res.xml");
-        new Gzip(new TestResource("repodata/primary.xml.gz.example").file()).unpack(res);
-        MatcherAssert.assertThat(
-            Files.readAllLines(res).toArray(),
-            Matchers.arrayContaining(
-                Files.readAllLines(
-                    new TestResource("repodata/primary.xml.example").file()
-                ).toArray()
-            )
-        );
+    /**
+     * Obtains resources from context loader.
+     * @return Path
+     */
+    public Path file() {
+        try {
+            return Paths.get(
+                Thread.currentThread().getContextClassLoader()
+                    .getResource(this.name).toURI()
+            );
+        } catch (final URISyntaxException ex) {
+            throw new IllegalStateException("Failed to load test recourses", ex);
+        }
     }
 
 }
